requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
flask>=2.0.0
pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
rasterio>=1.2.0
albumentations>=1.0.0
exifread>=2.3.0
pydantic>=1.8.0
pyyaml>=6.0
requests>=2.25.0
xgboost>=1.5.0